Add RT Info attributes deserialization
